update cron job

+++ src/partner/anhanguera/anhanguera.controller.ts
@@ -1,4 +1,9 @@
-import { Controller, Post, Headers } from '@nestjs/common';
+import {
+  Controller,
+  Post,
+  Headers,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { AnhangueraService } from './anhanguera.service';
 import { ApiExcludeController } from '@nestjs/swagger';
 
@@ -8,19 +13,31 @@ export class AnhangueraController {
   constructor(private readonly anhangueraService: AnhangueraService) {}
 
   @Post('sync-all')
-  async syncAll() {
-    const totalOffers = await this.anhangueraService.syncAllOffers();
+  syncAll(@Headers('x-api-key') apiKey: string): {
+    success: boolean;
+    message: string;
+  } {
+    if (apiKey !== process.env.CRON_API_KEY) {
+      throw new UnauthorizedException('Invalid API Key');
+    }
+
+    // ✅ Executa em background (não trava o n8n nem precisa de await)
+    void this.anhangueraService.syncAllOffers();
 
     return {
       success: true,
-      message: 'Sync finalizado com sucesso',
-      totalOffers,
+      message: 'Sincronização iniciada em segundo plano.',
     };
   }
 
   @Post('flush-all')
-  async flushAll() {
+  async flushAll(@Headers('x-api-key') apiKey: string) {
+    if (apiKey !== process.env.CRON_API_KEY) {
+      throw new UnauthorizedException('Invalid API Key');
+    }
+
     await this.anhangueraService.deleteAllAnhangueraData();
+
     return {
       success: true,
       message: 'Dados da Anhanguera removidos do Redis.',
